requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
gunicorn>=20.0,<21.0
djangorestframework==3.12.2
django-filter==2.4.0
Markdown==3.3.3
django-request==1.5.6